calback関連の謎のバグを修正、マルチページ構成に変更
